requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
cryptography==3.2.1
Django==3.1.5
django-celery-results==2.0.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
dnspython==1.16.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
eventlet==0.30.1
greenlet==1.0.0
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
numpy==1.20.1
openapi-codec==1.3.2
openpyxl==3.0.6
packaging==20.8
pandas==1.2.2
Pillow==8.1.0
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyshorteners==1.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
